serving/requirements.txt
torch==2.4.1
transformers==4.44.2
scikit-learn==1.5.2
joblib==1.4.2
flask==3.0.3
gunicorn==23.0.0
google-cloud-storage==2.18.2
loguru==0.7.2
